requirements.txt
# Core dependencies
streamlit==1.31.0
pandas==2.1.4
plotly==5.18.0
supabase==1.2.0
python-dotenv==1.0.0

# Additional dependencies for better date handling
python-dateutil>=2.8.2
pytz>=2023.3

# Data validation and processing
numpy>=1.24.0
openpyxl>=3.1.2  # For Excel file support
xlrd>=2.0.1      # For older Excel file support

# Visualization enhancements
plotly-express>=0.4.1

# Development tools (optional)
black  # Code formatting
flake8  # Code linting
pytest  # Testing

